+++ MansionMashers/MansionMashers/SourceFiles/Shop.c
@@ -120,7 +120,7 @@ void CreateWeaponShop(float shopPosX, float shopPosY, int newID, int collisionGr
 
 	//Weapon
 	newShop->Weapon = CreateDroppedWeapon(collisionGroup, weaponRarity, 250, 250, newID, shopPosX + 50, shopPosY);
-	newShop->Weapon->WeaponSprite->Rotation = (float)FOX_PI / 2;
+	newShop->Weapon->WeaponSprite->Rotation = FOX_PI / 2;
 	newShop->Weapon->CurrentShop = newShop;
 
 	//Gongratz
